project_1/pyproject.toml
[project]
name = "project-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chainlit>=2.5.5",
    "google-auth>=2.40.3",
    "litellm>=1.72.6.post1",
    "openai-agents>=0.0.18",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
]
